Add support for CDNSKEY record type
