+++ dns/rdtypes/ANY/CDNSKEY.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2003-2007, 2009-2011 Nominum, Inc.
+# Copyright (C) 2004-2007, 2009-2011 Nominum, Inc.
 #
 # Permission to use, copy, modify, and distribute this software and its
 # documentation for any purpose with or without fee is hereby granted,
@@ -13,35 +13,8 @@
 # ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT
 # OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
-"""Class ANY (generic) rdata type classes."""
+import dns.rdtypes.dnskeybase
+from dns.rdtypes.dnskeybase import *
 
-__all__ = [
-    'AFSDB',
-    'CDNSKEY',
-    'CDS',
-    'CERT',
-    'CNAME',
-    'DLV',
-    'DNAME',
-    'DNSKEY',
-    'DS',
-    'GPOS',
-    'HINFO',
-    'HIP',
-    'ISDN',
-    'LOC',
-    'MX',
-    'NS',
-    'NSEC',
-    'NSEC3',
-    'NSEC3PARAM',
-    'PTR',
-    'RP',
-    'RRSIG',
-    'RT',
-    'SOA',
-    'SPF',
-    'SSHFP',
-    'TXT',
-    'X25',
-]
+class CDNSKEY(dns.rdtypes.dnskeybase.DNSKEYBase):
+    """CDNSKEY record"""
